http_replay/requirements.txt
python-dotenv
azure-servicebus
httpx